requirements.txt
# ── Core framework ───────────────────────────────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# ── Document extraction ─────────────────────────────────────────────────────
PyMuPDF>=1.23.0
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.0.0

# ── NLP — spaCy (local NER) ─────────────────────────────────────────────────
spacy>=3.7.0

# ── AI — Groq (free cloud API for summarization + sentiment) ────────────────
groq>=0.4.0

# ── Testing ──────────────────────────────────────────────────────────────────
pytest>=7.4.0
httpx>=0.25.0
